refactor: import DockerContainer from module root
